Add files via upload
Card class is the generic, but should be able to start on GUI using the two.  Since the java generic card images are numbered 1-52 add 1 to the suit and multiply the suit by the value to get overall value within the deck. Also, card values are 1-13 which will need to be adjusted in the future.  Is this what you were needing? Any questions let me know.